- added min/max coords (AABB) in vertex stats (mesh) - made the space handedness configurable (LHS remains default) - added global rendering-mode overrides for some Object rendering modes. - added set_parameter() with an int, is_linked(), and get_id() to the   GfxProg class. - rearranged some camera code. - half-assed trackball implementation in scene_view example.

+++ 3dengfx/src/common/psort.hpp
@@ -21,7 +21,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  * Author: Mihalis Georgoulopoulos 2005
  */
 
-#include <string.h>
+#include <cstring>
 #include <algorithm>
 
 #ifndef _PSORT_HEADER_
@@ -50,7 +50,7 @@ template <class T, class P> void sort(T *elements, P *priorities, unsigned int n
 		criterion = less;
 
 	P **pointers = new P*[n];
-	for (unsigned long i=0; i<n; i++)
+	for (int i=0; i<n; i++)
 		pointers[i] = priorities + i;
 
 	// sort priority pointers
@@ -59,7 +59,7 @@ template <class T, class P> void sort(T *elements, P *priorities, unsigned int n
 	// collect sorted items
 	T *sorted_el = new T[n];
 	P *sorted_pr = new P[n];
-	for (unsigned long i=0; i<n; i++)
+	for (int i=0; i<n; i++)
 	{
 		sorted_el[i] = elements[pointers[i] - priorities];
 		sorted_pr[i] = *pointers[i];
